backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==4.45
fastapi==0.109.0
fastavro==1.9.3
frozenlist==1.4.1
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
yarl==1.9.4
zipp==3.17.0
